requirements.txt
boto3==1.21.27
Jinja2==2.11.3
markupsafe==2.0.1
google-cloud-bigquery==2.2.0
